+++ SilverSim/Database.MySQL/_Migration/MigrationElements.cs
@@ -415,7 +415,7 @@ namespace SilverSim.Database.MySQL._Migration
                 }
                 else if(f.IsEnum)
                 {
-                    def = (uint)Convert.ChangeType(def, f.GetEnumUnderlyingType());
+                    def = Convert.ChangeType(def, f.GetEnumUnderlyingType());
                 }
                 result.Add(colInfo.Name, string.Format("{0} {1} DEFAULT '{2}'",
                     typeSql,
